+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog File
 for Miss Stacy
 
+## v0.20
+- [x] Add plugin system
+- [x] Add jokes, memes plugins
+- [x] Add Spinal Cord
+
 ## v0.10
 
 - [x] main.js file now acts as a launcher for the interfaces